Implementation of the Graph Store Protocol
This includes HEAD, GET, PUT, POST, and DELETE.

In the case of POST, multipart/form-data is supported too.

For the POST method, there is one special parameter not part of the
specification: 'mimetype' can be used to override the mimetype.  If
not set, the Content-type header of each part is used, and if it is
also not present, the Content-type is guessed from the filename.

This implementation requires that a Dataset is used, ConjunctiveGraph
or Graph won't work.